+++ tests/smoke/test_smoke_checks.py
@@ -1,4 +1,4 @@
-from pyparsing import dbl_slash_comment
+# from pyparsing import dbl_slash_comment
 from tests.smoke.utils import open_dataset
 
 from owslib.wps import ComplexDataInput
@@ -26,6 +26,10 @@ C3S_CMIP6_MON_LEVEL_COLLECTION = (
     "c3s-cmip6.CMIP.CSIRO-ARCCSS.ACCESS-CM2.historical.r1i1p1f1.Amon.ta.gn.v20191108"
 )
 
+C3S_CMIP6_360DAY_CALENDAR_COLLECTION = (
+    "c3s-cmip6.ScenarioMIP.MOHC.HadGEM3-GC31-LL.ssp245.r1i1p1f3.day.pr.gn.v20190908"
+)
+
 C3S_CMIP5_DAY_COLLECTION = (
     "c3s-cmip5.output1.IPSL.IPSL-CM5B-LR.historical.day.atmos.day.r1i1p1.tas.v20120718"
 )
@@ -43,6 +47,12 @@ C3S_IPCC_ATLAS_CMIP5_COLLECTION = "c3s-ipcc-atlas.tnn.CMIP5.rcp45.mon"
 C3S_IPCC_ATLAS_CMIP6_COLLECTION = "c3s-ipcc-atlas.tnn.CMIP6.historical.mon"
 
 C3S_IPCC_ATLAS_CORDEX_COLLECTION = "c3s-ipcc-atlas.tnn.CORDEX-AFR.historical.mon"
+
+C3S_CICA_ATLAS_ERA5_COLLECTION = "c3s-cica-atlas.cd.ERA5.yr.v1"
+
+C3S_CICA_ATLAS_CORDEX_COLLECTION = "c3s-cica-atlas.cdd.CORDEX-CORE.historical.yr.v1"
+
+C3S_CICA_ATLAS_CMIP6_COLLECTION = "c3s-cica-atlas.cd.CMIP6.historical.yr.v1"
 
 WF_C3S_CMIP5 = json.dumps(
     {
@@ -80,6 +90,65 @@ WF_C3S_CMIP6 = json.dumps(
     }
 )
 
+WF_C3S_CMIP6_W_AVG = json.dumps(
+    {
+        "doc": "subset+weighted_average on cmip6",
+        "inputs": {"ds": [C3S_CMIP6_MON_COLLECTION]},
+        "outputs": {"output": "weighted_average/output"},
+        "steps": {
+            "subset": {
+                "run": "subset",
+                "in": {"collection": "inputs/ds", "time": "2020/2020"},
+            },
+            "weighted_average": {
+                "run": "weighted_average",
+                "in": {"collection": "subset/output"},
+            },
+        },
+    }
+)
+
+WF_C3S_CMIP6_REGRID = json.dumps(
+    {
+        "doc": "subset+regrid on cmip6",
+        "inputs": {"ds": [C3S_CMIP6_MON_COLLECTION]},
+        "outputs": {"output": "regrid/output"},
+        "steps": {
+            "subset": {
+                "run": "subset",
+                "in": {"collection": "inputs/ds", "time": "2016/2016"},
+            },
+            "regrid": {
+                "run": "regrid",
+                "in": {
+                    "collection": "subset/output",
+                    "method": "nearest_s2d",
+                    "grid": "1deg",
+                },
+            },
+        },
+    }
+)
+
+TC_ALL_DAYS = "day:01,02,03,04,05,06,07,08,09,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31"
+
+WF_C3S_CMIP6_360DAY_CALENDAR = json.dumps(
+    {
+        "doc": "subset on cmip6",
+        "inputs": {"ds": [C3S_CMIP6_360DAY_CALENDAR_COLLECTION]},
+        "outputs": {"output": "subset/output"},
+        "steps": {
+            "subset": {
+                "run": "subset",
+                "in": {
+                    "collection": "inputs/ds",
+                    "time": "2015/2015",
+                    "time_components": f"month:01,02,03|{TC_ALL_DAYS}",
+                },
+            },
+        },
+    }
+)
 
 WF_C3S_CORDEX = json.dumps(
     {
@@ -137,7 +206,7 @@ WF_C3S_CMIP6_DECADAL = json.dumps(
                 "run": "subset",
                 "in": {
                     "collection": "average/output",
-                    "time": "1995/1996",
+                    "time": "1995/1995",
                 },
             },
         },
@@ -252,6 +321,20 @@ def test_smoke_execute_c3s_cmip6_subset_metadata(wps, tmp_path):
     assert "coordinates" not in ds.time_bnds.encoding
 
 
+def test_smoke_execute_c3s_cmip6_subset_area_fill_value(wps, tmp_path):
+    dsid = "c3s-cmip6.CMIP.NCAR.CESM2-WACCM.historical.r1i1p1f1.day.tas.gn.v20190227"
+    inputs = [
+        ("collection", dsid),
+        ("time", "2010-01-01/2010-12-30"),
+        ("area", "5,49,7,51"),
+    ]
+    urls = wps.execute("subset", inputs)
+    assert len(urls) == 1
+    assert "tas_day_CESM2-WACCM_historical_r1i1p1f1_gn_20100101-20101230.nc" in urls[0]
+    ds = open_dataset(urls[0], tmp_path)
+    assert "tas" in ds.variables
+
+
 def test_smoke_execute_c3s_cordex_subset(wps, tmp_path):
     inputs = [
         ("collection", C3S_CORDEX_MON_COLLECTION),
@@ -293,6 +376,19 @@ def test_smoke_execute_c3s_cmip6_subset_by_point(wps, tmp_path):
     assert "rlds" in ds.variables
 
 
+def test_smoke_execute_c3s_cmip6_360calendar_subset_by_point(wps, tmp_path):
+    inputs = [
+        ("collection", C3S_CMIP6_360DAY_CALENDAR_COLLECTION),
+        ("time", "2015/2015"),
+        ("time_components", f"month:jan,feb,mar|{TC_ALL_DAYS}"),
+    ]
+    urls = wps.execute("subset", inputs)
+    assert len(urls) == 1
+    assert "pr_day_HadGEM3-GC31-LL_ssp245_r1i1p1f3_gn_20150101-20150330.nc" in urls[0]
+    ds = open_dataset(urls[0], tmp_path)
+    assert "pr" in ds.variables
+
+
 def test_smoke_execute_c3s_cordex_subset_by_point(wps, tmp_path):
     inputs = [
         ("collection", C3S_CORDEX_MON_COLLECTION),
@@ -318,18 +414,7 @@ def test_smoke_execute_c3s_cmip6_subset_original_files(wps):
     urls = wps.execute("subset", inputs)
     assert len(urls) == 1
     assert "data.mips.copernicus-climate.eu" in urls[0]
-
-
-# def test_smoke_execute_c3s_cmip5_subset_collection_only(wps):
-#     inputs = [("collection", C3S_CMIP5_MON_COLLECTION)]
-#     urls = wps.execute("subset", inputs)
-#     print(urls)
-#     assert len(urls) == 1
-#     assert "tas_mon_MPI-ESM-LR_historical_r1i1p1_18500116-20051216.nc" in urls[0]
-#     assert (
-#         "https://data.mips.copernicus-climate.eu/thredds/fileServer/esg_c3s-cmip5"
-#         in urls[0]
-#     )
+    assert "esg_c3s-cmip6" in urls[0]
 
 
 def test_smoke_execute_c3s_cmip6_subset_collection_only(wps):
@@ -348,10 +433,8 @@ def test_smoke_execute_c3s_cordex_subset_collection_only(wps):
     urls = wps.execute("subset", inputs)
     print(urls)
     assert len(urls) == 10
-    assert (
-        "https://data.mips.copernicus-climate.eu/thredds/fileServer/esg_c3s-cordex"
-        in urls[0]
-    )
+    assert "data.mips.copernicus-climate.eu" in urls[0]
+    assert "esg_c3s-cordex" in urls[0]
     assert (
         "tas_EUR-11_MOHC-HadGEM2-ES_rcp85_r1i1p1_CLMcom-CCLM4-8-17_v1_mon_200601-201012.nc"
         in urls[0]
@@ -435,6 +518,29 @@ def test_smoke_execute_c3s_cordex_average_time(wps):
     )
 
 
+def test_smoke_execute_c3s_cmip6_weighted_average(wps):
+    inputs = [("collection", C3S_CMIP6_MON_COLLECTION)]
+    urls = wps.execute("weighted_average", inputs)
+    assert len(urls) == 1
+    assert (
+        "rlds_Amon_INM-CM5-0_ssp245_r1i1p1f1_gr1_20150116-21001216_w-avg.nc" in urls[0]
+    )
+
+
+def test_smoke_execute_c3s_cmip6_regrid(wps):
+    inputs = [
+        ("collection", C3S_CMIP6_MON_COLLECTION),
+        ("grid", "auto"),
+        ("method", "nearest_s2d"),
+    ]
+    urls = wps.execute("regrid", inputs)
+    assert len(urls) == 1
+    assert (
+        "rlds_Amon_INM-CM5-0_ssp245_r1i1p1f1_gr_20150116-21001216_regrid-nearest_s2d-120x179_cells_grid.nc"
+        in urls[0]
+    )
+
+
 def test_smoke_execute_c3s_cmip5_orchestrate(wps):
     inputs = [
         ("workflow", ComplexDataInput(WF_C3S_CMIP5)),
@@ -451,6 +557,38 @@ def test_smoke_execute_c3s_cmip6_orchestrate(wps):
     urls = wps.execute("orchestrate", inputs)
     assert len(urls) == 1
     assert "rlds_Amon_INM-CM5-0_ssp245_r1i1p1f1_gr1_avg-t.nc" in urls[0]
+
+
+def test_smoke_execute_c3s_cmip6_weighted_average_orchestrate(wps):
+    inputs = [
+        ("workflow", ComplexDataInput(WF_C3S_CMIP6_W_AVG)),
+    ]
+    urls = wps.execute("orchestrate", inputs)
+    assert len(urls) == 1
+    assert (
+        "rlds_Amon_INM-CM5-0_ssp245_r1i1p1f1_gr1_20200116-20201216_w-avg.nc" in urls[0]
+    )
+
+
+def test_smoke_execute_c3s_cmip6_regrid_orchestrate(wps):
+    inputs = [
+        ("workflow", ComplexDataInput(WF_C3S_CMIP6_REGRID)),
+    ]
+    urls = wps.execute("orchestrate", inputs)
+    assert len(urls) == 1
+    assert (
+        "rlds_Amon_INM-CM5-0_ssp245_r1i1p1f1_gr_20160116-20161216_regrid-nearest_s2d-180x360_cells_grid.nc"
+        in urls[0]
+    )
+
+
+def test_smoke_execute_c3s_cmip6_360day_calendar_orchestrate(wps):
+    inputs = [
+        ("workflow", ComplexDataInput(WF_C3S_CMIP6_360DAY_CALENDAR)),
+    ]
+    urls = wps.execute("orchestrate", inputs)
+    assert len(urls) == 1
+    assert "pr_day_HadGEM3-GC31-LL_ssp245_r1i1p1f3_gn_20150101-20150330.nc" in urls[0]
 
 
 def test_smoke_execute_c3s_cmip6_orchestrate_metadata(wps, tmp_path):
@@ -495,40 +633,71 @@ def test_smoke_execute_c3s_cmip6_decadal_orchestrate(wps):
     ]
     urls = wps.execute("orchestrate", inputs)
     assert len(urls) == 1
-    assert (
-        "tas_Amon_HadGEM3-GC31-MM_dcppA-hindcast_r10i1p1f2_gn_19951116-19961216.nc"
-        in urls[0]
-    )
+    assert "tas_Amon_HadGEM3-GC31-MM_dcppA-hindcast" in urls[0]
+    assert "19951116-19951216.nc" in urls[0]
 
 
 def test_smoke_execute_c3s_ipcc_atlas_cmip5_subset(wps):
     inputs = [
         ("collection", C3S_IPCC_ATLAS_CMIP5_COLLECTION),
-        # ("time", "2000-01-01/2000-12-30"),
-        # ("original_files", "1"),
     ]
     urls = wps.execute("subset", inputs)
     assert len(urls) == 1
     assert "data.mips.copernicus-climate.eu" in urls[0]
+    assert "esg_c3s-ipcc-atlas" in urls[0]
+    assert "tnn_CMIP5_rcp45_mon_200601-210012.nc" in urls[0]
 
 
 def test_smoke_execute_c3s_ipcc_atlas_cmip6_subset(wps):
     inputs = [
         ("collection", C3S_IPCC_ATLAS_CMIP6_COLLECTION),
-        # ("time", "2000-01-01/2000-12-30"),
-        # ("original_files", "1"),
     ]
     urls = wps.execute("subset", inputs)
     assert len(urls) == 1
     assert "data.mips.copernicus-climate.eu" in urls[0]
+    assert "esg_c3s-ipcc-atlas" in urls[0]
+    assert "tnn_CMIP6_historical_mon_185001-201412.nc" in urls[0]
 
 
 def test_smoke_execute_c3s_ipcc_atlas_cordex_subset(wps):
     inputs = [
         ("collection", C3S_IPCC_ATLAS_CORDEX_COLLECTION),
-        # ("time", "2000-01-01/2000-12-30"),
-        # ("original_files", "1"),
     ]
     urls = wps.execute("subset", inputs)
     assert len(urls) == 1
     assert "data.mips.copernicus-climate.eu" in urls[0]
+    assert "esg_c3s-ipcc-atlas" in urls[0]
+    assert "tnn_CORDEX-AFR_historical_mon_197001-200512.nc" in urls[0]
+
+
+def test_smoke_execute_c3s_cica_atlas_cmip6_subset(wps):
+    inputs = [
+        ("collection", C3S_CICA_ATLAS_CMIP6_COLLECTION),
+        ("time", "2000/2000"),
+    ]
+    urls = wps.execute("subset", inputs)
+    assert len(urls) == 1
+    assert "data.mips.copernicus-climate.eu" not in urls[0]
+    assert "cd_CMIP6_historical_yr_20000101-20000101.nc" in urls[0]
+
+
+def test_smoke_execute_c3s_cica_atlas_cordex_subset(wps):
+    inputs = [
+        ("collection", C3S_CICA_ATLAS_CORDEX_COLLECTION),
+        ("time", "2000/2000"),
+    ]
+    urls = wps.execute("subset", inputs)
+    assert len(urls) == 1
+    assert "data.mips.copernicus-climate.eu" not in urls[0]
+    assert "cdd_CORDEX-CORE_historical_yr_20000101-20000101.nc" in urls[0]
+
+
+def test_smoke_execute_c3s_cica_atlas_era5_subset_no_time_param(wps):
+    inputs = [
+        ("collection", C3S_CICA_ATLAS_ERA5_COLLECTION),
+    ]
+    urls = wps.execute("subset", inputs)
+    assert len(urls) == 1
+    assert "data.mips.copernicus-climate.eu" in urls[0]
+    assert "esg_c3s-cica-atlas" in urls[0]
+    assert "cd_ERA5_yr_1940-2022.nc" in urls[0]