+++ tests/test_workflow.py
@@ -42,27 +42,12 @@ def test_run_tree_wf(tmp_path):
     assert "tas_mon_HadGEM2-ES_rcp85_r1i1p1_20850101-21200101_avg-year.nc" in output[0]
 
 
-def test_run_tree_wf_2(tmp_path):
-    wf = workflow.WorkflowRunner(output_dir=tmp_path)
-    output = wf.run(TREE_WF_2)
-    assert "tas_mon_HadGEM2-ES_rcp85_r1i1p1_20900101-21000101_avg-year.nc" in output[0]
-
-
 def test_run_wf_cmip6_subset_average(tmp_path):
     wfdoc = resource_file("wf_cmip6_subset_average.json")
     wf = workflow.WorkflowRunner(output_dir=tmp_path)
     output = wf.run(wfdoc)
     assert (
         "rlds_Amon_IPSL-CM6A-LR_historical_r1i1p1f1_gr_19850101-20140101_avg-year.nc"
-        in output[0]
-    )
-
-
-def test_run_tree_wf_5(tmp_path):
-    wf = workflow.WorkflowRunner(output_dir=tmp_path)
-    output = wf.run(TREE_WF_5)
-    assert (
-        "rlds_Amon_IPSL-CM6A-LR_historical_r1i1p1f1_gr_19950101-20000101_avg-year.nc"
         in output[0]
     )